+++ README.md
@@ -1,19 +1,3 @@
-# ![1676383092268](image/README/1676383092268.png) Color picker
+# ![1675328739763](image/README/1675328739763.png) Color picker
 
-With my color picker, you can :
-
-* Copy/paste color picked in hex and rgb
-* Save some color, if you want (10 max)
-* and choose color with the picker or pick a color
-
-## Preview
-
-![1676383341549](image/README/1676383341549.png)
-
-## How to install
-
-1) Open your extensions manager tab (`<browser>://extensions`)
-2) Enable "Developper Mode"
-3) Drag and drop the downloaded zip file into the tab
-4) Once uploaded and loaded, disable "Developper Mode"
-5) Have fun using "Color Picker" ! 😸
+Just simply pick color of your choice !
