+++ app/assets/CSS/style.css
@@ -7,25 +7,16 @@
     --c-main: #212D40;
     --c-second: #364156;
     --c-color1: #7D4E57;
-    --c-color2: #D66853;
-    --c-tooltip: #242424;
+    --c-color2: #D66853; 
 }
-body{background-color: var(--c-bg);border-radius: 5px;padding: 5px;width: 500px;height: auto;}
+body{background-color: var(--c-bg);border-radius: 5px;padding: 5px;width: 382px;}
 .logo{
     float: left;
     position: relative;
     height: 48px;width: auto;
-    padding-top: 5px;
-    padding-left: 5px;
+    padding-top: 1rem;
+    padding-left: 1rem;
 }
-.pick{
-    position: absolute;
-    margin-top: -85%;
-    margin-left: 18%;
-    text-align: center;
-    pointer-events: none;
-}
-p{margin:0}h1{margin:5px}
 .titre{
     color:whitesmoke;
     width: 100%;text-align: center;
@@ -39,28 +30,8 @@ p{margin:0}h1{margin:5px}
     border-radius: 5px;
     border: 1px solid var(--c-second);
     box-shadow: 0px 1px 10px -3px var(--c-second);
-    padding: 0.5rem;color: whitesmoke;
+    padding: 1rem;color: whitesmoke;
     font-size: large;
-}
-.hex:hover{
-    background-color: var(--c-bg);
-    color: whitesmoke;
-    cursor: pointer;
-    transition: 400ms;
-}
-input[type="button"]{
-    background-color: var(--c-second);
-    padding: 0.5rem;border-radius: 5px;
-    border: none;width: 5rem;
-    color: whitesmoke;
-}
-input[type="button"]:hover{
-    background-color: var(--c-bg);
-    padding: 0.5rem;border-radius: 5px;
-    border: none;width: 5rem;
-    color: whitesmoke;
-    cursor: pointer;
-    transition: 400ms;
 }
 input[type="color"] {
     border-radius: 5px;
@@ -127,30 +98,47 @@ input[type="color"]::-webkit-color-swatch {
     -ms-flex-item-align: auto;
     align-self: auto;
 }
-.center{margin-left: auto;margin-right: auto;}
-#result-rgb, #result-hex{text-align: center;margin-left: 0px;}
+.center{
+    margin-left: auto;
+    margin-right: auto;
+}
+#result-rgb, #result-hex{
+    text-align: center;
+    margin-left: 0px;
+}
 .red,.green,.blue,.hex{
     background-color: var(--c-second);
     padding: 0.5rem;border-radius: 5px;
     margin: 5px;
 }
-.red:hover,.green:hover,.blue:hover{
-    background-color: var(--c-bg);
-    padding: 0.5rem;border-radius: 5px;
-    margin: 5px;transition: 400ms;
-    cursor:pointer;
+.tooltip {
+    position: relative;
+    display: inline-block;
 }
-.clr{
-    background-Color:whitesmoke;
-    border-Radius:5px;
-    height:3rem;width:3.5rem;
-    float:Left;position:relative;
-    margin-Left:5px;margin-Top:5px;
-    cursor:not-allowed;
+.tooltip .tooltiptext {
+    visibility: hidden;
+    width: 140px;
+    background-color: #555;
+    color: #fff;
+    text-align: center;
+    border-radius: 6px;
+    padding: 5px;
+    position: absolute;
+    z-index: 1;
+    bottom: 150%;
+    left: 50%;
+    margin-left: -75px;
+    opacity: 0;
+    transition: opacity 0.3s;
 }
-#history{
-    padding: 1rem;height: 130px;
-    margin: 0 1.5rem 0 1.5rem;
+.tooltip .tooltiptext::after {
+    content: "";
+    position: absolute;
+    top: 100%;
+    left: 50%;
+    margin-left: -5px;
+    border-width: 5px;
+    border-style: solid;
+    border-color: #555 transparent transparent transparent;
 }
-.color-pick{min-width: 70%; padding-top: 1rem}
-.color-main{min-width:30%}
+.tooltip:hover .tooltiptext {visibility: visible;opacity: 1;}
