+++ manifest.json
@@ -1,7 +1,7 @@
 {
   "manifest_version": 3,
   "name": "Color Picker",
-  "description": "Select the color that you wish !",
+  "description": "Just simply pick color of your choice !",
   "version": "1.0",
   "action": {
     "default_title": "Click to view a popup",
@@ -11,9 +11,7 @@
     "16": "/app/assets/icons/icon-16.png",
     "32": "/app/assets/icons/icon-32.png",
     "48": "/app/assets/icons/icon-48.png",
-    "128": "/app/assets/icons/icon-128.png",
-    "256": "/app/assets/icons/icon-256.png",
-    "512": "/app/assets/icons/icon-512.png"
+    "128": "/app/assets/icons/icon-128.png"
   },
   "background": [{
     "service_worker": "/app/assets/js/core.js"
